Added tests

+++ src/test/java/com/springapp/utils/user/UserStubFactory.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * This is a factory class that contains stub for User fo tetst.
+ * This is a factory class that contains stub for User fo test.
  */
 public class UserStubFactory {
 
@@ -23,7 +23,7 @@ public class UserStubFactory {
     public static final List<User> createStubbedUserList(final int numOfUser) {
         final List<User> list = new ArrayList<>();
         for (int index = 0; index < numOfUser; index++) {
-            list.add(createStubbedUser(Long.valueOf(index)));
+            list.add(createStubbedUser(index));
         }
         return list;
     }
@@ -34,7 +34,7 @@ public class UserStubFactory {
      * @param userId
      * @return
      */
-    public static final User createStubbedUser(final Long userId) {
+    public static final User createStubbedUser(final Integer userId) {
         final User user = new User();
         user.setId(userId);
         user.setUsername("some_username");
@@ -48,7 +48,7 @@ public class UserStubFactory {
      * @param username
      * @return
      */
-    public static final User createStubUserWithCustomFields(final Long userId,
+    public static final User createStubUserWithCustomFields(final Integer userId,
                                                             final String username) {
         final User user = new User();
         user.setUsername(username);
